chore: move directory Avatar
