+++ frontend/index.html
@@ -29,7 +29,7 @@
 	<link rel="stylesheet" href="./src/components/FriendGameRecords/FriendOneGameRecord.css">
 	<link rel="stylesheet" href="./src/components/FriendGameRecords/FriendOutcome.css">
 	<link rel="stylesheet" href="./src/components/FriendSearchModal/FriendSearchModal.css">
-	<link rel="stylesheet" href="./src/components/Profile/Avatar.css">
+	<link rel="stylesheet" href="./src/components/Avatar/Avatar.css">
 </head>
 
 <body>
