+++ frontend/src/components/Avatar/Avatar.js
